library: move rwlock into subdir
